snap-sync: Allow to print usage as non-root user
This patch allows us to use '-h' as non-root user.

At the moment, running snap-sync always fails in case that the script is
started at non-root user.

There is no reason, why normal users should not be allowed to print the
help screen.

Additionally, this patch allows to print the help screen, even if the
snapper config is not present.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>