snap-sync: fix error formatting bug
If snap-sync is invoked with an erroneous cmdline parameter, it will
print something like:

$ snap-sync -ahjk                                                                                                                                                                1
==> ERROR: Unknown option: -ahjk\nRun 'snap-sync -h' for valid options.

Control characters are not resolved. Reason is the erroneous usage of
printf(), as %s format specifiers do not resolve control characters.

Simply use echo, and redirect it to stderr. Output of notify_error shall
not go anywhere in case of errors to prevent production of additional
garbage.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>